backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
PyMySQL==1.1.0
mysqlclient==2.2.0
django-environ==0.21.0
gunicorn==21.2.0
whitenoise==6.6.0
